+++ modules/pytket-quantinuum/tests/api_test.py
@@ -12,7 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import Tuple
+from io import StringIO
+from typing import Any, Tuple
 
 from requests_mock.mocker import Mocker
 
@@ -27,17 +28,17 @@ def test_quum_login(
     """Test that credentials are storable and deletable using
     the QuantinuumQAPI handler."""
 
-    username, pwd = mock_credentials
+    _, pwd = mock_credentials
 
     # Check credentials are retrievable
-    assert mock_quum_api_handler._cred_store.login_credential(username) == pwd
+    assert mock_quum_api_handler._cred_store._password == pwd
     assert mock_quum_api_handler._cred_store.refresh_token == mock_token
     assert mock_quum_api_handler._cred_store.id_token == mock_token
 
     # Delete authentication and verify
     mock_quum_api_handler.delete_authentication()
     assert mock_quum_api_handler._cred_store.id_token == None
-    assert mock_quum_api_handler._cred_store.login_credential(username) == None
+    assert mock_quum_api_handler._cred_store._password == None
     assert mock_quum_api_handler._cred_store.refresh_token == None
 
 
@@ -47,7 +48,7 @@ def test_machine_status(
 ) -> None:
     """Test that we can retrieve the machine state via  Quantinuum endpoint."""
 
-    machine_name = "HQS-LT-S1-APIVAL"
+    machine_name = "quum-LT-S1-APIVAL"
     mock_machine_state = "online"
 
     mock_url = f"https://qapi.quantinuum.com/v1/machine/{machine_name}"
@@ -61,5 +62,55 @@ def test_machine_status(
 
     assert mock_quum_api_handler.status(machine_name) == mock_machine_state
 
-    # Delete authentication tokens to clean them from the keyring
+    # Delete authentication tokens to clean them from memory
     mock_quum_api_handler.delete_authentication()
+
+
+def test_full_login(
+    requests_mock: Mocker,
+    mock_credentials: Tuple[str, str],
+    mock_token: str,
+    monkeypatch: Any,
+) -> None:
+    username, pwd = mock_credentials
+
+    mock_url = "https://qapi.quantinuum.com/v1/login"
+
+    requests_mock.register_uri(
+        "POST",
+        mock_url,
+        json={
+            "id-token": mock_token,
+            "refresh-token": "refresh" + mock_token,
+        },
+        headers={"Content-Type": "application/json"},
+    )
+
+    # fake user input from stdin
+    monkeypatch.setattr("sys.stdin", StringIO(username + "\n"))
+    monkeypatch.setattr("getpass.getpass", lambda prompt: pwd)
+
+    api_handler = QuantinuumQAPI()
+    # emulate no pytket config stored email address
+    api_handler.config.username = None
+    api_handler.full_login()
+
+    assert api_handler._cred_store.id_token == mock_token
+    assert api_handler._cred_store.refresh_token == "refresh" + mock_token
+    assert api_handler._cred_store._id_token_timeout is not None
+    assert api_handler._cred_store._refresh_token_timeout is not None
+
+    assert api_handler._cred_store._password is None
+    assert api_handler._cred_store._user_name is None
+
+    api_handler.delete_authentication()
+
+    assert all(
+        val is None
+        for val in (
+            api_handler._cred_store.id_token,
+            api_handler._cred_store.refresh_token,
+            api_handler._cred_store._id_token_timeout,
+            api_handler._cred_store._refresh_token_timeout,
+        )
+    )